Completed interactive pricing, and added pricing / day to gears
