+++ app/views/bookings/new.html.erb
@@ -10,9 +10,5 @@
   <%= f.input :end_time, label: "End Date", as: :string, input_html: { class: "datepicker" }, wrapper_html: { class: "w-5 inline_field_wrapper end_date" } %>
   <%= f.submit "Book", class: "btn btn-primary" %>
 <% end %>
-<div class="container inline-flex">
-  <p id="error"></p>
-  <p id="total"></p>
-  <p id="price"></p>
-  <p id="days"></p>
-<div>
+
+<p id="total"></p>
